requirements.txt
# Valid Person Finder - Backend
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
ddgs>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
# Bonus: Agentic pipeline (CrewAI)
crewai>=0.80.0
litellm
